Modularity for Existing Functions Part 1
